requirements-nlp.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
spacy==3.8.7
yake==0.4.8
pandas==2.2.3
scikit-learn==1.5.1
xgboost==2.1.1
lightgbm==4.5.0
joblib==1.4.2
psycopg[binary]==3.2.9
sentence-transformers==3.0.1
